+++ Sources/SwiftUIKit/Views/Form/FormTextField.swift
@@ -29,11 +29,45 @@ public struct FormTextField: View {
     ) {
         self.title = title
         self.placeholder = placeholder
+        self.isMultiline = false
+        self.lineLimit = 1
+        self.reservesSpace = false
+        self._text = text
+    }
+    
+    /**
+     Create a form multiline text editor.
+
+     - Parameters:
+       - title: The view title.
+       - placeholder: The text field placeholder text.
+       - isMultiline: Whether or not the text field is multiline.
+       - lineLimit: An optional line limit, by default `1`.
+       - reservesSpace: Whether or not to reserve the line space, by default `false`.
+       - text: The text to edit.
+     */
+    @available(iOS 16.0, *)
+    public init(
+        title: String,
+        placeholder: String,
+        isMultiline: Bool,
+        lineLimit: Int = 1,
+        reservesSpace: Bool = false,
+        text: Binding<String>
+    ) {
+        self.title = title
+        self.placeholder = placeholder
+        self.isMultiline = isMultiline
+        self.lineLimit = lineLimit
+        self.reservesSpace = reservesSpace
         self._text = text
     }
     
     private let title: String
     private let placeholder: String
+    private let isMultiline: Bool
+    private let lineLimit: Int
+    private let reservesSpace: Bool
     
     @Binding
     private var text: String
@@ -41,20 +75,77 @@ public struct FormTextField: View {
     public var body: some View {
         VStack(alignment: .leading, spacing: 4) {
             FormTextTitle(title)
-            TextField(placeholder, text: $text)
+            if #available(iOS 16.0, *), isMultiline {
+                TextField(
+                    placeholder,
+                    text: $text,
+                    axis: .vertical
+                ).tryApplyLines(lineLimit, reserve: reservesSpace)
+            } else {
+                TextField(
+                    placeholder,
+                    text: $text
+                )
+            }
         }.padding(.vertical, 3)
     }
 }
 
+@available(iOS 16.0, *)
+private extension View {
+    
+    @ViewBuilder
+    func tryApplyLines(_ lines: Int, reserve: Bool) -> some View {
+        if lines > 1 {
+            self.lineLimit(lines, reservesSpace: reserve)
+        } else {
+            self
+        }
+    }
+}
+
+@available(iOS 16.0, *)
 struct FormTextField_Previews: PreviewProvider {
     
-    static var previews: some View {
-        List {
-            FormTextField(
-                title: "Title",
-                placeholder: "Enter text",
-                text: .constant("Text value"))
+    struct Preview: View {
+        
+        @State
+        private var text = ""
+        
+        var body: some View {
+            List {
+                FormTextField(
+                    title: "Title",
+                    placeholder: "Enter text",
+                    text: $text
+                )
+                FormTextField(
+                    title: "Title",
+                    placeholder: "Enter text",
+                    isMultiline: true,
+                    text: $text
+                )
+                FormTextField(
+                    title: "Title",
+                    placeholder: "Enter text",
+                    isMultiline: true,
+                    lineLimit: 5,
+                    text: $text
+                )
+                FormTextField(
+                    title: "Title",
+                    placeholder: "Enter text",
+                    isMultiline: true,
+                    lineLimit: 5,
+                    reservesSpace: true,
+                    text: $text
+                )
+            }
         }
+    }
+    
+    static var previews: some View {
+        Preview()
     }
 }
 #endif